Fix Panel not responding properly to focus change
Because window advertisements use yutani_wait_for, we need to actually
loop over yutani_poll() until the queue reports that it is empty,
otherwise, we'll fall into fswait - and since we've already read (and
buffered) some calls we haven't processed, nothing will happen until
either another event is delivered or the timeout occurs. This was
causing the panel to sometimes just not respond to a click because the
click was delivered during window advertisements, so we wouldn't process
it ever, since we didn't read the queue on a timeout.