Fix some more mouse issues.
This mostly affects VirtualBox with mouse integration disabled, as QEMU
never sends mouse changes big enough for this to be a problem, but we
just were not handling the 9-bit x/y changes. That's what we get for
going from five years of QEMU to a year of VirtualBox with mouse
integration... Anyway, it's all working quite nicely now!