QEMU display harness
This is a combination of a harness for running QEMU and tracking its
window size, and a client running in the ToaruOS guest to receive those
window size updates and perform mode setting. Run the harness with
`util/qemu-harness.py` and then run `qemu_display_hack.py` from the
guest (using Alt-F2, preferrably) and the display will automatically
set the appropriate mode (and send ctrl-alt-u to qemu to disable
scaling) whenever the window changes. This is an experimental hack, it
only works on Linux (uses Xlib to do the window size querying), only
works with the SDL output (GTK both doesn't like it when you try to
resize the window too small, and it has that darn menu bar that makes it
hard to size everything correctly since that's theme-specific), and is
very experimental. That said, it performs quite nicely, possibly even
better than the native support in VirtualBox.