New graphical login screen.
Fixes #143

The new login screen runs separately from the actual login session
manager and communicates back with a pipe. This may be totally insecure
if you dig into it, but keep in mind our password hash policy is just an
unsalted sha and our sudo doesn't care who you are, so all in all I
don't think it's a problem at this stage. Maybe we can do something
better in the future.

The login screen is a Python app and features some new design updates:
the panel is available (with a modified set of widgets). The actual
login boxes are in their own window, which can be moved around, but this
is mostly to improve performance by allowing that window to be updated
repeatedly.

I might add more panel widgets in the future, such as accessibility
options; having access to things like IME and on-screen keyboard
settings would definitely be useful at the login prompt.